🔥 Restructure search page
